Implementación sistema de autenticación para Login y protección de URLs con Spring Security

+++ src/main/java/com/udla/proyectofin_web_mvc/modelo/Producto.java
@@ -31,9 +31,12 @@ public class Producto {
     }
 
     // Getters y Setters
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
-
+    public Long getId() {
+        return id;
+    }
+    public void setId(Long id) {
+        this.id = id;
+    }
     public String getNombre() { return nombre; }
     public void setNombre(String nombre) { this.nombre = nombre; }
